fix(trusty-search): split finish.rs teardown helpers to clear the 500-SLOC cap
Main's combination of #5514 and #5527 (each fine alone against their own
base) pushed crates/trusty-search/src/service/reindex/finish.rs to 501 SLOC,
tripping the required "500-line file-size cap" check on every subsequent
PR's head — a second red-main condition of the same class as the
non_exhaustive break fixed earlier in this PR.

Extracts the four post-loop teardown/completion helpers (stop_pollers,
resolve_corpus_swap, resolve_hnsw_swap, rebuild_kg) — everything finish.rs's
own module doc already describes as a separable seam — into a new sibling
module finish_teardown.rs. finish_reindex and its two context structs
(BatchTotals, FinishCtx) stay in finish.rs. No logic changed; pure move.

bash scripts/check_line_cap.sh: 0 violations (was 1)
cargo fmt --check -p trusty-search: clean
cargo clippy -p trusty-search --all-targets -- -D warnings: clean
cargo test -p trusty-search --lib reindex::: 148 passed; 0 failed
